requirements.txt
fastapi
uvicorn
pydantic
rdkit
numpy
pandas
torch-geometric
torch-scatter
torch-sparse